Re-factored the run browser into its own sub-package within Firefly.
