+++ pytest.ini
@@ -5,3 +5,4 @@ asyncio_mode = auto
 testpaths =
   src/haven/tests
   src/firefly/tests
+  src/firefly/run_browser/tests
