+++ src/firefly/run_browser/tests/test_client.py
@@ -1,0 +1,49 @@
+import pytest
+
+
+from firefly.run_browser.client import DatabaseWorker
+
+
+@pytest.mark.asyncio
+async def test_filter_runs(catalog):
+    worker = DatabaseWorker(catalog=catalog)
+    runs = await worker.load_all_runs(filters={"plan": "xafs_scan"})
+    # Check that the runs were filtered
+    assert len(runs) == 1
+
+
+@pytest.mark.asyncio
+async def test_distinct_fields(catalog):
+    worker = DatabaseWorker(catalog=catalog)
+    distinct_fields = await worker.load_distinct_fields()
+    # Check that the dictionary has the right structure
+    for key in ["sample_name"]:
+        assert key in distinct_fields.keys()
+    
+
+
+# -----------------------------------------------------------------------------
+# :author:    Mark Wolfman
+# :email:     wolfman@anl.gov
+# :copyright: Copyright © 2023, UChicago Argonne, LLC
+#
+# Distributed under the terms of the 3-Clause BSD License
+#
+# The full license is in the file LICENSE, distributed with this software.
+#
+# DISCLAIMER
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+# HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+# SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+# LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+# DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+# THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+# -----------------------------------------------------------------------------
+
